feat(contract): changes in update discounts , markups , supplements

+++ src/app/pages/contract/update-contract/season-details/supplement/supplement.component.ts
@@ -3,6 +3,7 @@ import { FormBuilder, FormGroup, ReactiveFormsModule, Validators } from '@angula
 import { ActivatedRoute, Router } from '@angular/router';
 import { CommonModule } from '@angular/common';
 import { SupplementService } from '../../../../../services/supplement/supplement.service';
+import { AlertService } from '../../../../../services/alert/alert.service';
 
 @Component({
   selector: 'app-supplement',
@@ -20,6 +21,7 @@ export class SupplementComponent implements OnInit {
     private fb: FormBuilder,
     private route: ActivatedRoute,
     private router: Router,
+    private alertService: AlertService,
     private supplementService: SupplementService
   ) {
     this.supplementForm = this.fb.group({
@@ -65,25 +67,30 @@ export class SupplementComponent implements OnInit {
 
   async onSubmitSupplementDetails(supplement: any) {
     if (supplement.form.invalid) {
+      this.alertService.showError('All fields are required');
       return;
     }
     const token: string | null = localStorage.getItem('token');
+
     if (!token) {
       console.error('Token not found');
       return;
     }
     try {
       console.log('Supplement Form Data:', supplement.form.value);
-      const response = await this.supplementService.updateSupplement(supplement.id, supplement.form.value, token).toPromise();
+      const response = await this.supplementService.updateSupplement(supplement.supplementID, supplement.form.value, token).toPromise();
       console.log('Supplement updated:', response);
+      this.alertService.showSuccess('Supplement Updated Successfully');
       this.loadSupplements();
     } catch (error: any) {
+      this.alertService.showError('Error occurred while updating Supplement');
       console.error('Error updating supplement:', error);
     }
   }
 
   async onAddNewSupplement() {
     if (this.newSupplementForm.invalid) {
+      this.alertService.showError('All fields are required');
       return;
     }
     const token: string | null = localStorage.getItem('token');
@@ -95,25 +102,40 @@ export class SupplementComponent implements OnInit {
       console.log('New Supplement Form Data:', this.newSupplementForm.value);
       const response = await this.supplementService.addSupplementToSeason(this.seasonID, this.newSupplementForm.value, token).toPromise();
       console.log('New Supplement added:', response);
+      this.alertService.showSuccess('supplement Added Successfully');
       this.loadSupplements();
       this.newSupplementForm.reset();
     } catch (error: any) {
+      this.alertService.showError('Error occurred while Adding supplement');
       console.error('Error adding new supplement:', error);
     }
   }
 
   async onDeleteSupplement(supplementId: number) {
+
+    const confirmDelete = await this.alertService.showConfirm(
+      'Are you sure you want to Delete this Discount?',
+      'Do you want to proceed?',
+      'Yes, proceed',
+      'No, cancel'
+    );
+
+    if (!confirmDelete) return;
+
     const token: string | null = localStorage.getItem('token');
     if (!token) {
       console.error('Token not found');
       return;
     }
-    try {
-      const response = await this.supplementService.deleteSupplement(supplementId, token).toPromise();
-      console.log('Supplement deleted:', response);
-      this.loadSupplements();
-    } catch (error: any) {
-      console.error('Error deleting supplement:', error);
+    
+    if(confirmDelete){
+      try {
+        const response = await this.supplementService.deleteSupplement(supplementId, token).toPromise();
+        console.log('Supplement deleted:', response);
+        this.loadSupplements();
+      } catch (error: any) {
+        console.error('Error deleting supplement:', error);
+      }
     }
   }
 }
